+++ test/bigtest/interactors/job-profiles/index.js
@@ -1,3 +1,0 @@
-export * from './job-profiles';
-export * from './job-profile-form';
-export * from './job-profile-details';