+++ test/bigtest/interactors/mapping-profiles/index.js
@@ -1,3 +1,0 @@
-export * from './mapping-profiles';
-export * from './mapping-profile-details';
-export * from './mapping-profiles-form';