+++ test/bigtest/interactors/match-profiles/index.js
@@ -1,3 +1,0 @@
-export * from './match-profiles';
-export * from './match-profile-form';
-export * from './match-profile-details';