+++ test/bigtest/network/factories/file.js
@@ -1,8 +1,0 @@
-import { Factory } from 'miragejs';
-import faker from 'faker';
-
-export default Factory.extend({
-  id() {
-    return faker.random.uuid();
-  },
-});