+++ test/bigtest/network/models/file.js
@@ -1,8 +1,0 @@
-// models
-
-import {
-  Model,
-  belongsTo,
-} from 'miragejs';
-
-export default Model.extend({ uploadDefinition: belongsTo() });