+++ test/bigtest/network/scenarios/fetch-job-log-success.js
@@ -1,8 +1,0 @@
-import { jobLog } from '../../mocks/job-log';
-
-export default server => {
-  server.get('/source-storage/records', {
-    records: jobLog,
-    totalRecords: jobLog.length,
-  });
-};