+++ test/bigtest/network/scenarios/fetch-jobs-data-success.js
@@ -1,8 +1,0 @@
-import { jobExecutions } from '../../mocks';
-
-export default server => {
-  server.get('/metadata-provider/jobExecutions', {
-    jobExecutions,
-    totalRecords: jobExecutions.length,
-  });
-};