+++ test/bigtest/network/scenarios/tags-enabled.js
@@ -1,4 +1,0 @@
-export default server => server.get(
-  '/configurations/entries',
-  { configs: [{ value: 'true' }] },
-);