+++ test/bigtest/network/scenarios/upload-definition-without-files.js
@@ -1,5 +1,0 @@
-export default server => {
-  server.create('upload-definition', { fileDefinitions: [] });
-  server.get('/data-import/uploadDefinitions');
-  server.delete('/data-import/uploadDefinitions/:id');
-};