+++ test/bigtest/network/serializers/file.js
@@ -1,3 +1,0 @@
-import { Serializer } from 'miragejs';
-
-export default Serializer.extend({ serializeIds: 'always' });